resubmission function

+++ Presentation/xDC.Web/Scripts/app/amsd/index.js
@@ -92,26 +92,27 @@
                                 return (e.row.data.formStatus == "Draft" && e.row.data.isFormOwner);
                             },
                             onClick: function (e) {
-                                var data = {
-                                    id: e.row.data.id
-                                };
+                                if (!confirm("Do you really want to delete this?")) {
+                                    return false;
+                                } else {
+                                    var data = {
+                                        id: e.row.data.id
+                                    };
 
-                                $.ajax({
-                                    type: "delete",
-                                    url: '../api/amsd/DeleteInflowFundDraftForm',
-                                    data: data,
-                                    success: function (data) {
-                                        $("#error_container").bs_success("Draft deleted");
-                                        $amsdGrid.refresh();
-                                    },
-                                    fail: function(jqXHR, textStatus, errorThrown) {
-                                        $("#error_container").bs_alert(textStatus + ': ' + errorThrown);
-                                    }
-                                });
-                                e.event.preventDefault();
-
-
-                                e.event.preventDefault();
+                                    $.ajax({
+                                        type: "delete",
+                                        url: '../api/amsd/DeleteInflowFundDraftForm',
+                                        data: data,
+                                        success: function (data) {
+                                            $("#error_container").bs_success("Draft deleted");
+                                            $amsdGrid.refresh();
+                                        },
+                                        fail: function (jqXHR, textStatus, errorThrown) {
+                                            $("#error_container").bs_alert(textStatus + ': ' + errorThrown);
+                                        }
+                                    });
+                                    e.event.preventDefault();
+                                }
                             }
                         },
                         {
